requirements.txt
langchain-google-genai
requests
python-dotenv
langchain_google_genai
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0